pyproject.toml
[tool.poetry]
name = "search-spotify-music"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
spotipy = "^2.24.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
polars = "^1.1.0"
ruff = "^0.5.2"
numpy = "^2.0.0"
pip-licenses = "^4.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
